TOSQUASH: move `Subplace` trait to `ops::place`
