+++ library/core/src/field.rs
@@ -1,7 +1,6 @@
 //! Field Reflection
 
 use crate::marker::PhantomData;
-use crate::ptr::Pointee;
 
 /// Field Representing Type
 #[unstable(feature = "field_representing_type_raw", issue = "none")]
@@ -84,36 +83,4 @@ pub unsafe trait Field: Send + Sync + Copy {
     /// The offset of the field in bytes.
     #[lang = "field_offset"]
     const OFFSET: usize = crate::intrinsics::field_offset::<Self>();
-}
-
-/// A subplace of [`Self::Source`] with the type [`Self::Target`].
-///
-/// A subplace is always within the same allocation as the base place. Current subplaces are:
-/// - field accesses,
-/// - array/slice indexes.
-///
-/// This represents an arbitrary chaining of these; it can also be empty.
-///
-/// # Safety
-///
-/// FIXME
-#[unstable(feature = "field_projections", issue = "145383")]
-#[rustc_deny_explicit_impl]
-#[rustc_dyn_incompatible_trait]
-#[lang = "subplace"]
-pub unsafe trait Subplace: Sized + Copy {
-    /// The type of the base place this subplace is a part of.
-    #[lang = "subplace_source"]
-    type Source: ?Sized;
-
-    /// The type of this subplace.
-    #[lang = "subplace_target"]
-    type Target: ?Sized;
-
-    /// The offset of this subplace.
-    #[lang = "subplace_offset"]
-    fn offset(
-        self,
-        metadata: <Self::Source as Pointee>::Metadata,
-    ) -> (usize, <Self::Target as Pointee>::Metadata);
 }
